current state fpr latest esp-idf

+++ components/arduino-esp32/cores/esp32/esp32-hal-uart.c
@@ -18,10 +18,10 @@
 #include "freertos/task.h"
 #include "freertos/queue.h"
 #include "freertos/semphr.h"
-#include "rom/ets_sys.h"
+#include "esp32/rom/ets_sys.h"
 #include "esp_attr.h"
-#include "esp_intr.h"
-#include "rom/uart.h"
+#include "esp_intr_alloc.h"
+#include "esp32/rom/uart.h"
 #include "soc/uart_reg.h"
 #include "soc/uart_struct.h"
 #include "soc/io_mux_reg.h"
@@ -253,7 +253,7 @@ size_t uartResizeRxBuffer(uart_t * uart, size_t new_size) {
         vQueueDelete(uart->queue);
         uart->queue = xQueueCreate(new_size, sizeof(uint8_t));
         if(uart->queue == NULL) {
-            return NULL;
+            return (size_t)NULL;
         }
     }
     UART_MUTEX_UNLOCK();
